Added remaining modifications.

+++ 16-Oct-2025/src/avgn.c
@@ -4,23 +4,23 @@
 #include <stdio.h>
 
 int main() {
-	size_t n;
+	int n;
 	double avg;
 	printf("Enter the size of the array: ");
-	scanf("%zu", &n);
+	scanf("%d", &n);
 	int data[n], sum = 0;
 
 	// Separated input and processing for demonstration.
-	for (size_t i = 0; i < n; i++) {
-		printf("Enter number to store in data[%zu]: ", i);
+	for (int i = 0; i < n; i++) {
+		printf("Enter number to store in data[%d]: ", i);
 		scanf("%d", &data[i]);
 	}
 
-	for (size_t i = 0; i < n; i++)
+	for (int i = 0; i < n; i++)
 		sum += data[i];
 
 	avg = (double) sum / n;
 
-	printf("Average of %zu numbers = %.2lf\n", n, avg);
+	printf("Average of %d numbers = %.2lf\n", n, avg);
 	return 0;
 }
